+++ CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+### Added
+- Add request info into not found response
+
 ### Fixed
 - Remove default value for `LOCALHOST` environment value in info
 - Disable deprecated PHP warnings in face API